requirements.txt
cherrypy
cp_sqlalchemy
passlib
jinja2
alembic
